+++ modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/dto/v2_0/util/MapToDDMFormValuesConverterUtil.java
@@ -14,6 +14,9 @@
 
 package com.liferay.data.engine.rest.internal.dto.v2_0.util;
 
+import com.liferay.data.engine.rest.internal.strategy.DefaultMapToDDMFormValuesConverterStrategy;
+import com.liferay.data.engine.rest.internal.strategy.MapToDDMFormValuesConverterStrategy;
+import com.liferay.data.engine.rest.internal.strategy.NestedFieldsSupportMapToDDMFormValuesConverterStrategy;
 import com.liferay.dynamic.data.mapping.model.DDMForm;
 import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
 import com.liferay.portal.kernel.util.MapUtil;
